folder distribution
