+++ app/src/main/java/com/homies/homies/RecyclerViewAdapterListUser.java
@@ -9,7 +9,7 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.homies.homies.services.AddUserGroupResponse;
+import com.homies.homies.retrofit.model.group.AddUserGroupResponse;
 
 import java.util.ArrayList;
