+++ app/src/main/java/com/homies/homies/retrofit/model/GroupRequest.java
@@ -1,4 +1,4 @@
-package com.homies.homies.services;
+package com.homies.homies.retrofit.model;
 
 import android.service.autofill.UserData;
