+++ app/src/main/java/com/homies/homies/retrofit/model/UserData.java
@@ -1,4 +1,4 @@
-package com.homies.homies.services;
+package com.homies.homies.retrofit.model;
 
 import java.util.List;
