+++ app/src/main/java/com/homies/homies/retrofit/model/group/AddUserGroupRequest.java
@@ -1,6 +1,6 @@
-package com.homies.homies.services;
+package com.homies.homies.retrofit.model.group;
 
-public class DeleteUserGroupRequest {
+public class AddUserGroupRequest {
 
     private int idAdminGroup;
     private String login;
