--- app/src/main/java/com/homies/homies/services/TaskList.java
@@ -1,24 +1,0 @@
-package com.homies.homies.services;
-
-public class TaskList {
-
-    private int id;
-    private String nameList;
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getNameList() {
-        return nameList;
-    }
-
-    public void setNameList(String nameList) {
-        this.nameList = nameList;
-    }
-}
-
