+++ app/src/main/java/com/homies/homies/user/LoginFragment.java
@@ -1,6 +1,5 @@
 package com.homies.homies.user;
 
-import android.annotation.SuppressLint;
 import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
@@ -26,9 +25,9 @@ import com.google.android.material.bottomsheet.BottomSheetDialog;
 import com.homies.homies.MenuActivity;
 import com.homies.homies.R;
 import com.homies.homies.services.Adaptador;
-import com.homies.homies.services.ApiClient;
-import com.homies.homies.services.UserRequest;
-import com.homies.homies.services.UserResponse;
+import com.homies.homies.retrofit.config.NetworkConfig;
+import com.homies.homies.retrofit.model.UserRequest;
+import com.homies.homies.retrofit.model.UserResponse;
 
 import retrofit2.Call;
 import retrofit2.Callback;
@@ -114,7 +113,7 @@ public class LoginFragment extends Fragment {
     }
 
     public void loginUser(UserRequest userRequest) {
-        Call<UserResponse> loginResponseCall = ApiClient.getService().loginUser(userRequest);
+        Call<UserResponse> loginResponseCall = NetworkConfig.getService().loginUser(userRequest);
         loginResponseCall.enqueue(new Callback<UserResponse>() {
             @Override
             public void onResponse(Call<UserResponse> call, Response<UserResponse> response) {
@@ -141,7 +140,7 @@ public class LoginFragment extends Fragment {
     }
 
     public void resetPassword(String email) {
-        Call<Void> resetPasswordResponseCall = ApiClient.getService().resetPassword(email);
+        Call<Void> resetPasswordResponseCall = NetworkConfig.getService().resetPassword(email);
         resetPasswordResponseCall.enqueue(new Callback<Void>() {
             @Override
             public void onResponse(Call<Void> call, Response<Void> response) {
